implemented fire for all enemies and added wave/level spawning

+++ SmallGameProject/SmallGameProject/Ability.h
@@ -16,7 +16,7 @@ protected:
 
 	float arc;
 	int nrOfProjectiles;
-	int cooldown;
+	float cooldown;
 	int attackDelay;
 	int maxCharges;
 	int triggerDelay;
@@ -29,7 +29,7 @@ protected:
 public:
 	Ability();
 	virtual~Ability();
-	virtual void Initialize(float arc, int nrOfProjectiles, int cooldown, int attackDelay, int maxCharges, int triggerDelay);
+	virtual void Initialize(float arc, int nrOfProjectiles, float cooldown, int attackDelay, int maxCharges, int triggerDelay);
 	virtual float activate(Enemy* enemy, EntitySubject* entitySubject, DirectX::XMFLOAT3 playerPos) = 0;
 	virtual void update(Enemy* enemy, EntitySubject* entitySubject, float deltaTime);
